requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.14
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.4.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.8
psycopg2-binary==2.9.6
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.2
style==1.1.0
tqdm==4.65.0
typing_extensions==4.7.1
tzlocal==5.0.1
update==0.0.1
urllib3==2.0.3
Werkzeug==2.2.2
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
